refactor(engine): extract event writer interface
