+++ engine/src/main/java/io/zeebe/engine/processing/streamprocessor/writers/TypedStreamWriter.java
@@ -11,11 +11,11 @@ import io.zeebe.engine.processing.streamprocessor.TypedRecord;
 import io.zeebe.msgpack.UnpackedObject;
 import io.zeebe.protocol.impl.record.RecordMetadata;
 import io.zeebe.protocol.record.RejectionType;
-import io.zeebe.protocol.record.intent.Intent;
 import java.util.function.Consumer;
 
 /** Things that only a stream processor should write to the log stream (+ commands) */
-public interface TypedStreamWriter extends TypedCommandWriter {
+public interface TypedStreamWriter extends TypedCommandWriter, TypedEventWriter {
+
   void appendRejection(
       TypedRecord<? extends UnpackedObject> command, RejectionType type, String reason);
 
@@ -24,13 +24,6 @@ public interface TypedStreamWriter extends TypedCommandWriter {
       RejectionType type,
       String reason,
       Consumer<RecordMetadata> metadata);
-
-  void appendNewEvent(long key, Intent intent, UnpackedObject value);
-
-  void appendFollowUpEvent(long key, Intent intent, UnpackedObject value);
-
-  void appendFollowUpEvent(
-      long key, Intent intent, UnpackedObject value, Consumer<RecordMetadata> metadata);
 
   void configureSourceContext(long sourceRecordPosition);
 }
